+++ tests/test_scripts.py
@@ -1,18 +1,14 @@
 import unittest
 import unittest.mock
 
-from tutor.commands import config as tutor_config
+from tutor import config as tutor_config
 from tutor import scripts
 
 
-class DummyRunner(scripts.BaseRunner):
-    exec = unittest.mock.Mock()
-
-
 class ScriptsTests(unittest.TestCase):
-    def test_run(self):
-        config = tutor_config.load_defaults()
-        runner = DummyRunner("/tmp", config)
-        rendered_script = runner.render("create_databases.sh")
-        runner.run("someservice", "create_databases.sh")
-        runner.exec.assert_called_once_with("someservice", rendered_script)
+    def test_is_activated(self):
+        config = {"ACTIVATE_SERVICE1": True, "ACTIVATE_SERVICE2": False}
+        runner = scripts.BaseRunner("/tmp", config)
+
+        self.assertTrue(runner.is_activated("service1"))
+        self.assertFalse(runner.is_activated("service2"))