+++ tutor/templates/apps/openedx/settings/cms/common.py
@@ -28,3 +28,6 @@ LOCALE_PATHS.append("/openedx/locale")
 for folder in [LOG_DIR, MEDIA_ROOT, STATIC_ROOT_BASE]:
     if not os.path.exists(folder):
         os.makedirs(folder)
+
+{{ patch("openedx-common-settings") }}
+{{ patch("openedx-cms-common-settings") }}